% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BX2way.r
\docType{data}
\name{BX2way}
\alias{BX2way}
\title{Book-crossing 2-way contingency table data}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(BX2way)
}
\description{
30x15 contingency table classifying book reviews by book ISBN (30 levels) and country of reviewer (15 levels).
The dataset appears as a list prepared for contingency table resampling conditional on the marginal totals.
Components of the list are A (the configuration matrix); y (marginal totals); x (true cell counts); lambda (cell means under independence model); and MarkovBasis (a matrix with columns corresponding
to the vectors in full Markov basis for the resampling problem).
}
\examples{
data(BX2way)
Xsampler(A=BX2way$A,y=BX2way$y,lambda=BX2way$lambda,Model="Uniform",Method="Gibbs",tune.par=0.5,combine=FALSE,burnin=50,ndraws=200)
}
\keyword{datasets}
