% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.r
\name{effectiveSize}
\alias{effectiveSize}
\title{Effective sample size}
\usage{
effectiveSize(x)
}
\arguments{
\item{x}{A matrix of MCMC out, with rows correspondings to variables and columns to iterations.}
}
\value{
Effective sample sizes for each variable
}
\description{
Version of effectiveSize from coda package, modified for ease of use with DynamicLatticeBasis package.
}
\examples{
data(BXno3way)
res <- Xsampler(A=BXno3way$A,y=BXno3way$y,lambda=BXno3way$lambda,Model="Poisson",Method="Gibbs",tune.par=0.5,combine=TRUE)
effectiveSize(res$X)
}
