% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-circuit.r
\name{is_circuit}
\alias{is_circuit}
\title{Check whether basis vectors are circuits}
\usage{
is_circuit(A, U)
}
\arguments{
\item{A}{Configuration matrix with ncol(A) >= nrow(A)}

\item{U}{A matrix with columns comprising some basis for the integer lattice kernel of A}
}
\value{
A logical vector of length ncol(U) indicating whether each column of U is a circuit
}
\description{
A circuit is a vector with minimal support by inclusion and coprime (non-zero) entries. This function checks whether each column of a matrix U is a circuit of the solution set for y = Ax
}
