% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-wholenumber.r
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Checking for integer values}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
Returns TRUE if all entries of x are whole numbers; FALSE otherwise
}
\description{
This function tests whether all entries in a vector are integers (up to rounding error)
}
\examples{
is_wholenumber(c(1,2,4.1))
}
