% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-redundancy.r
\name{remove_redundancy}
\alias{remove_redundancy}
\title{Remove redundancy frrom linear system}
\usage{
remove_redundancy(A, y)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Vector of observed data. Should be of length nrow(A)}
}
\value{
A list with components A and y, respectively the configuration matrix and data vector with redundant rows removed.
}
\description{
Removes redundant rows of configuration matric and corresponding entries in observed data
}
