% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontends.r
\name{XMarkovBasis_MH}
\alias{XMarkovBasis_MH}
\title{Front ends for Xsampler}
\usage{
XMarkovBasis_MH(
  y,
  A,
  lambda,
  U,
  x.ini = NULL,
  Model = "Poisson",
  Proposal = "NonUnif",
  NB.alpha = 0,
  ndraws = 10000,
  burnin = 2000,
  verbose = 0,
  THIN = 1
)
}
\arguments{
\item{...}{As for Xsampler}
}
\value{
A list as for Xsampler
}
\description{
Front ends for Xsampler, for quick implementation of dynamic lattice basis and full Markov basis samplers. XGibbs and XMH are Gibbs and Metropolis-Hastings versions of Xsampler using dynamic lattice bases. XMarkovBasis_Gibbs and XMarkovBasis_MH are versions using a full Markov basis.
}
\examples{
data(BXno3way)
XGibbs(BXno3way$y,BXno3way$A,BXno3way$lambda,combine=T,tune.par=0.5)
}
