% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-sampler.r
\name{Xsampler}
\alias{Xsampler}
\title{Z-polytope sampler for linear inverse problems}
\usage{
Xsampler(
  y,
  A,
  lambda,
  U = NULL,
  Method = "MH",
  Reorder = TRUE,
  tune.par = 0.5,
  combine = FALSE,
  x.order = NULL,
  x.ini = NULL,
  Model = "Poisson",
  Proposal = "Unif",
  NB.alpha = 1,
  ndraws = 10000,
  burnin = 2000,
  verbose = 0,
  THIN = 1
)
}
\arguments{
\item{y}{Vector of observed count data.}

\item{A}{Model configuration matrix, assumed to be binary.}

\item{lambda}{Mean vector for x.}

\item{U}{Optional matrix the columns of which should be a Markov (sub)-basis.}

\item{Method}{"MH" for Metropolis-Hastings sampler, "Gibbs" for Gibbs sampler.}

\item{Reorder}{Should the columns of A be reordered? Defaults to TRUE.}

\item{tune.par}{Tuning parameter (alpha) controlling variation in fitness values for lattice bases. Defaults to 0.5.}

\item{combine}{Should extra moves be included combining lattice basis vectors? Defaults to FALSE, but should usually be set to TRUE if A is not unimodular.}

\item{x.order}{If Reorder=FALSE, x.order can be used to reorder columns of A to match ordering of entries of x. Defaults to NULL when no such reordering is performed.}

\item{x.ini}{Vector of initial values for x. Default is NULL, when initial values derived through integer programming.}

\item{Model}{"Poisson", "Uniform", "NegBin" or "Normal" (the last being a discrete approximation).}

\item{Proposal}{"NonUnif" or "Unif" (default).}

\item{NB.alpha}{Dispersion parameter for negaqtive-binomial distribution. Defaults to 1.}

\item{ndraws}{Number of iterations to run sampler after burn-in. One iteration comprises cycling through the full basis (possibly augmented by a combined move). Defaults to 10^4.}

\item{burnin}{Number of iteractions for burn in period. Defaults to 2000, which is usually more than adequate.}

\item{verbose}{Controls level of detail in recording lattice bases used.}

\item{THIN}{Thinning parameter for output. Defaults to 1 (no thinning).}
}
\value{
A list with components X (a matrix, each row corresponding to samplers for an entry of x) and x.order (a vector describing dynamic selection of lattice bases, if verbose=1).
}
\description{
For the linear inverse problem y=Ax where y and x are counts, the y-fibre (solution set) is a Z-polytope (i.e. the points on the integer lattice within a convex polytope). This function implements samplers for the Z-polytope, with using a dynamic lattice basis or a full Markov basis. The underyling model for x can be Poisson, uniform or negative binomial.
}
\examples{
data(LondonRoad)
Xsampler(A=LondonRoad$A,y=LondonRoad$y,lambda=LondonRoad$lambda,Model="Poisson",Method="Gibbs",tune.par=0.5,combine=FALSE)
}
